extract makeDraggable() function out from Cloud.js to Board.js
