+++ src/components/Board.js
@@ -5,6 +5,7 @@ import FlowText from './FlowText'
 import Cloud from './Cloud'
 import Arrow from './Arrow'
 import Parameter from './Parameter'
+import firebase from '../firebase'
 
 const boardStyle = {
     height: 1000,
@@ -19,6 +20,72 @@ const svgWrapper = {
 
 const markerId = "arrow"
 
+const updateCloudPosition = (cloudByFlow, x, y, _FB_PATH) => {    
+    firebase.database().ref(_FB_PATH+'state').once('value').then((state) => {
+        let targetCloud    
+        const cloudsOrigin = Object.assign([], state.val().cloudsOrigin)
+        const cloudsDestination = Object.assign([], state.val().cloudsDestination)
+        if(cloudsOrigin.find(cloud => cloud.flow === cloudByFlow)){
+            targetCloud = cloudsOrigin.find(cloud => cloud.flow === cloudByFlow)
+        } else {
+            targetCloud = cloudsDestination.find(cloud => cloud.flow === cloudByFlow)
+        }
+        targetCloud.posX = x
+        targetCloud.posY = y
+        firebase.database().ref(_FB_PATH+'state/cloudsOrigin').set(cloudsOrigin);
+        firebase.database().ref(_FB_PATH+'state/cloudsDestination').set(cloudsDestination)
+    })
+}
+
+const makeDraggable = (id, updatePositionHandler, _FB_PATH) => {    
+    let draggable = document.getElementById(id);
+    let offset
+
+    const mouseDown = (e) => {
+        draggable = document.getElementById(id);
+        offset = getMousePosition(e);
+        offset.x -= parseFloat(draggable.getAttributeNS(null, "x"));
+        offset.y -= parseFloat(draggable.getAttributeNS(null, "y"));
+        draggable.addEventListener('mousemove', mouseMove)
+    }
+
+    const mouseMove = (e) => {
+        if (!draggable) return
+        e.preventDefault();
+        const coord = getMousePosition(e);
+        updatePositionHandler(id, coord.x - offset.x, coord.y - offset.y, _FB_PATH)
+    }
+
+    const mouseUpOrLeave = (e) => {
+        draggable = null
+    }
+
+    // mouse position to svg position
+    const getMousePosition = (e) => {
+        const CTM = draggable.getScreenCTM();
+        return {
+            x: (e.clientX - CTM.e) / CTM.a,
+            y: (e.clientY - CTM.f) / CTM.d
+        };
+    }
+
+    draggable.addEventListener('mousedown', mouseDown)
+    draggable.addEventListener('mouseup', mouseUpOrLeave)
+    draggable.addEventListener('mouseleave', mouseUpOrLeave)
+}
+
+const DraggableComponent = (Component, updatePositionHandler) => 
+    class extends React.Component {
+        componentDidMount() {
+            makeDraggable(this.props.elementId, updatePositionHandler, this.props._FB_PATH)
+        }
+        render() {
+            return <Component {...this.props}/>
+        }
+    }
+
+const DraggableCloud = DraggableComponent(Cloud, updateCloudPosition)
+
 export default class Board extends React.Component {    
     
     render() {
@@ -32,7 +99,9 @@ export default class Board extends React.Component {
         })
 
         const cloudsOrigin = this.props.cloudsOrigin.map( cloud => {
-            return <Cloud
+            return <DraggableCloud
+                elementId={cloud.flow}
+                _FB_PATH={this.props._FB_PATH}
                 key={cloud.flow}
                 flow={cloud.flow}
                 posX={cloud.posX}
@@ -42,7 +111,8 @@ export default class Board extends React.Component {
         })
 
         const cloudsDestination = this.props.cloudsDestination.map( cloud => {
-            return <Cloud
+            return <DraggableCloud
+                elementId={cloud.flow}
                 key={cloud.flow}
                 flow={cloud.flow}
                 posX={cloud.posX}
